Extract camera/transform dialog cluster into engine/src/transformDialogs.ts
Moves showTranslationDialog, showQuaternionDialog, showAngleAxisDialog,
showCameraPositionDialog, showCameraRotationDialog, and
showRotationCenterDialog out of main.ts as thin wrappers, and drops the
now-unused dialog-template/matrix-helper imports.
