+++ engine/src/main.ts
@@ -44,25 +44,8 @@ import { registerDefaultReaders, readDepth } from './depth/DepthRegistry';
 import { normalizeDepth, projectToPointCloud } from './depth/DepthProjector';
 import { ColorImageLoader } from './colorImageLoader';
 import { PerfTimer, perfLog, setPerfSink } from './utils/perfLog';
-import {
-  createRotationMatrix,
-  createTranslationMatrix,
-  createQuaternionMatrix,
-  createAngleAxisMatrix,
-  parseSpaceSeparatedValues,
-  parseMatrixInput,
-} from './utils/matrix';
-import {
-  escapeHtml,
-  addTooltipsToTruncatedFilenames,
-  createModalDialog,
-  translationDialogTemplate,
-  quaternionDialogTemplate,
-  angleAxisDialogTemplate,
-  cameraPositionDialogTemplate,
-  cameraRotationDialogTemplate,
-  rotationCenterDialogTemplate,
-} from './ui/dialogs';
+import { createRotationMatrix, parseMatrixInput } from './utils/matrix';
+import { escapeHtml, addTooltipsToTruncatedFilenames } from './ui/dialogs';
 import * as sequencePlayback from './sequencePlayback';
 import * as pose from './pose';
 import * as renderStats from './renderStats';
@@ -88,6 +71,7 @@ import * as axesFeature from './axesFeature';
 import * as transformationMatrix from './transformationMatrix';
 import * as depthCameraParamsPrompt from './depthCameraParamsPrompt';
 import * as formatDataHandlers from './formatDataHandlers';
+import * as transformDialogs from './transformDialogs';
 import { formatFileSize } from './utils/format';
 import { ColorProcessor } from './colorProcessor';
 import { DepthConverter } from './depth/DepthConverter';
@@ -1369,15 +1353,15 @@ class PointCloudVisualizer {
     transformationMatrix.resetTransformationMatrix(this, fileIndex);
   }
 
-  private multiplyTransformationMatrices(fileIndex: number, matrix: THREE.Matrix4): void {
+  multiplyTransformationMatrices(fileIndex: number, matrix: THREE.Matrix4): void {
     transformationMatrix.multiplyTransformationMatrices(this, fileIndex, matrix);
   }
 
-  private addTranslationToMatrix(fileIndex: number, x: number, y: number, z: number): void {
+  addTranslationToMatrix(fileIndex: number, x: number, y: number, z: number): void {
     transformationMatrix.addTranslationToMatrix(this, fileIndex, x, y, z);
   }
 
-  private updateMatrixTextarea(fileIndex: number): void {
+  updateMatrixTextarea(fileIndex: number): void {
     transformationMatrix.updateMatrixTextarea(this, fileIndex);
   }
 
@@ -1385,7 +1369,7 @@ class PointCloudVisualizer {
     transformationMatrix.updateCameraMatrixDisplay(this);
   }
 
-  private updateCameraControlsPanel(): void {
+  updateCameraControlsPanel(): void {
     transformationMatrix.updateCameraControlsPanel(this);
   }
 
@@ -7100,323 +7084,27 @@ class PointCloudVisualizer {
   }
 
   private showTranslationDialog(fileIndex: number): void {
-    const { dialog, close } = createModalDialog(translationDialogTemplate());
-
-    const cancelBtn = dialog.querySelector('#cancel-translation');
-    const applyBtn = dialog.querySelector('#apply-translation');
-
-    if (cancelBtn) {
-      cancelBtn.addEventListener('click', close);
-    }
-
-    if (applyBtn) {
-      applyBtn.addEventListener('click', () => {
-        const input = (dialog.querySelector('#translation-input') as HTMLTextAreaElement).value;
-        const values = parseSpaceSeparatedValues(input);
-
-        if (values.length === 3) {
-          const [x, y, z] = values;
-          this.addTranslationToMatrix(fileIndex, x, y, z);
-          this.updateMatrixTextarea(fileIndex);
-          close();
-        } else {
-          alert('Please enter exactly 3 numbers for translation (X Y Z)');
-        }
-      });
-    }
+    transformDialogs.showTranslationDialog(this, fileIndex);
   }
 
   private showQuaternionDialog(fileIndex: number): void {
-    const { dialog, close } = createModalDialog(quaternionDialogTemplate());
-
-    const cancelBtn = dialog.querySelector('#cancel-quaternion');
-    const applyBtn = dialog.querySelector('#apply-quaternion');
-
-    if (cancelBtn) {
-      cancelBtn.addEventListener('click', close);
-    }
-
-    if (applyBtn) {
-      applyBtn.addEventListener('click', () => {
-        const input = (dialog.querySelector('#quaternion-input') as HTMLTextAreaElement).value;
-        const values = parseSpaceSeparatedValues(input);
-
-        if (values.length === 4) {
-          const [x, y, z, w] = values;
-          const quaternionMatrix = createQuaternionMatrix(x, y, z, w);
-          this.multiplyTransformationMatrices(fileIndex, quaternionMatrix);
-          this.updateMatrixTextarea(fileIndex);
-          close();
-        } else {
-          alert('Please enter exactly 4 numbers for the quaternion (X Y Z W)');
-        }
-      });
-    }
+    transformDialogs.showQuaternionDialog(this, fileIndex);
   }
 
   private showAngleAxisDialog(fileIndex: number): void {
-    const { dialog, close } = createModalDialog(angleAxisDialogTemplate());
-
-    const cancelBtn = dialog.querySelector('#cancel-angle-axis');
-    const applyBtn = dialog.querySelector('#apply-angle-axis');
-
-    if (cancelBtn) {
-      cancelBtn.addEventListener('click', close);
-    }
-
-    if (applyBtn) {
-      applyBtn.addEventListener('click', () => {
-        const input = (dialog.querySelector('#angle-axis-input') as HTMLTextAreaElement).value;
-        const values = parseSpaceSeparatedValues(input);
-
-        if (values.length === 4) {
-          const [axisX, axisY, axisZ, angleDegrees] = values;
-          const axis = new THREE.Vector3(axisX, axisY, axisZ);
-          const angle = (angleDegrees * Math.PI) / 180; // Convert to radians
-          const angleAxisMatrix = createAngleAxisMatrix(axis, angle);
-          this.multiplyTransformationMatrices(fileIndex, angleAxisMatrix);
-          this.updateMatrixTextarea(fileIndex);
-          close();
-        } else {
-          alert('Please enter exactly 4 numbers for axis and angle (X Y Z angle in degrees)');
-        }
-      });
-    }
+    transformDialogs.showAngleAxisDialog(this, fileIndex);
   }
 
   private showCameraPositionDialog(): void {
-    const currentPos = this.camera.position;
-    const { dialog, close } = createModalDialog(
-      cameraPositionDialogTemplate(
-        currentPos.x.toFixed(3),
-        currentPos.y.toFixed(3),
-        currentPos.z.toFixed(3)
-      )
-    );
-
-    const cancelBtn = dialog.querySelector('#cancel-camera-pos');
-    const applyBtn = dialog.querySelector('#apply-camera-pos');
-    const setAllZeroBtn = dialog.querySelector('#set-all-pos-zero');
-
-    if (cancelBtn) {
-      cancelBtn.addEventListener('click', close);
-    }
-
-    if (setAllZeroBtn) {
-      setAllZeroBtn.addEventListener('click', () => {
-        (dialog.querySelector('#camera-position-input') as HTMLTextAreaElement).value = '0 0 0';
-      });
-    }
-
-    if (applyBtn) {
-      applyBtn.addEventListener('click', () => {
-        const input = (dialog.querySelector('#camera-position-input') as HTMLTextAreaElement).value;
-        const constraint = (
-          dialog.querySelector('input[name="position-constraint"]:checked') as HTMLInputElement
-        ).value;
-        const values = parseSpaceSeparatedValues(input);
-
-        if (values.length === 3) {
-          const [x, y, z] = values;
-
-          // Store current camera state
-          const currentQuaternion = this.camera.quaternion.clone();
-          const currentTarget = this.controls.target.clone();
-
-          // Update position
-          this.camera.position.set(x, y, z);
-
-          // Apply constraint logic
-          if (constraint === 'rotation') {
-            // Keep rotation (angle) - restore quaternion
-            this.camera.quaternion.copy(currentQuaternion);
-
-            // Update target based on new position and preserved rotation
-            const direction = new THREE.Vector3(0, 0, -1);
-            direction.applyQuaternion(currentQuaternion);
-            this.controls.target.copy(this.camera.position.clone().add(direction));
-          } else {
-            // Keep rotation center (target) - restore target (default behavior)
-            this.controls.target.copy(currentTarget);
-
-            // Adjust camera rotation to look at the preserved target
-            this.camera.lookAt(currentTarget);
-          }
-
-          this.controls.update();
-          this.updateCameraControlsPanel();
-          close();
-        } else {
-          alert('Please enter exactly 3 numbers for position (X Y Z)');
-        }
-      });
-    }
+    transformDialogs.showCameraPositionDialog(this);
   }
 
   private showCameraRotationDialog(): void {
-    // Get rotation from quaternion to handle all camera operations consistently
-    const euler = new THREE.Euler();
-    euler.setFromQuaternion(this.camera.quaternion, 'XYZ');
-    const rotX = (euler.x * 180) / Math.PI;
-    const rotY = (euler.y * 180) / Math.PI;
-    const rotZ = (euler.z * 180) / Math.PI;
-
-    const { dialog, close } = createModalDialog(
-      cameraRotationDialogTemplate(rotX.toFixed(1), rotY.toFixed(1), rotZ.toFixed(1))
-    );
-
-    const cancelBtn = dialog.querySelector('#cancel-camera-rot');
-    const applyBtn = dialog.querySelector('#apply-camera-rot');
-    const setAllZeroBtn = dialog.querySelector('#set-all-rot-zero');
-
-    if (cancelBtn) {
-      cancelBtn.addEventListener('click', close);
-    }
-
-    if (setAllZeroBtn) {
-      setAllZeroBtn.addEventListener('click', () => {
-        (dialog.querySelector('#camera-rotation-input') as HTMLTextAreaElement).value = '0 0 0';
-      });
-    }
-
-    if (applyBtn) {
-      applyBtn.addEventListener('click', () => {
-        const input = (dialog.querySelector('#camera-rotation-input') as HTMLTextAreaElement).value;
-        const constraint = (
-          dialog.querySelector('input[name="rotation-constraint"]:checked') as HTMLInputElement
-        ).value;
-        const values = parseSpaceSeparatedValues(input);
-
-        if (values.length === 3) {
-          const [x, y, z] = values;
-
-          // Store current camera state
-          const currentPosition = this.camera.position.clone();
-          const currentTarget = this.controls.target.clone();
-
-          // Create quaternion from Euler angles
-          const rotationEuler = new THREE.Euler(
-            (x * Math.PI) / 180,
-            (y * Math.PI) / 180,
-            (z * Math.PI) / 180,
-            'XYZ'
-          );
-          const quaternion = new THREE.Quaternion();
-          quaternion.setFromEuler(rotationEuler);
-
-          // Apply constraint logic
-          if (constraint === 'position') {
-            // Keep position - restore position and apply rotation directly
-            this.camera.position.copy(currentPosition);
-            this.camera.quaternion.copy(quaternion);
-
-            // Update target based on new rotation and preserved position
-            const direction = new THREE.Vector3(0, 0, -1);
-            direction.applyQuaternion(quaternion);
-            this.controls.target.copy(this.camera.position.clone().add(direction));
-          } else {
-            // Keep rotation center - restore target and adjust position (default behavior)
-            const distance = currentPosition.distanceTo(currentTarget);
-            this.controls.target.copy(currentTarget);
-
-            // Position camera relative to preserved target
-            const direction = new THREE.Vector3(0, 0, distance);
-            direction.applyQuaternion(quaternion);
-            this.camera.position.copy(currentTarget).add(direction);
-
-            // Set up vector and look at target
-            const up = new THREE.Vector3(0, 1, 0);
-            up.applyQuaternion(quaternion);
-            this.camera.up.copy(up);
-            this.camera.lookAt(currentTarget);
-          }
-
-          this.controls.update();
-          this.updateCameraControlsPanel();
-          close();
-        } else {
-          alert('Please enter exactly 3 numbers for rotation (X Y Z degrees)');
-        }
-      });
-    }
+    transformDialogs.showCameraRotationDialog(this);
   }
 
   private showRotationCenterDialog(): void {
-    // Get current rotation center (controls target)
-    const target = this.controls.target;
-    const { dialog, close } = createModalDialog(
-      rotationCenterDialogTemplate(target.x.toFixed(3), target.y.toFixed(3), target.z.toFixed(3))
-    );
-
-    // Event listeners
-    const setOriginBtn = dialog.querySelector('#set-center-origin');
-    const cancelBtn = dialog.querySelector('#cancel-rotation-center');
-    const applyBtn = dialog.querySelector('#apply-rotation-center');
-
-    if (setOriginBtn) {
-      setOriginBtn.addEventListener('click', () => {
-        (dialog.querySelector('#rotation-center-input') as HTMLTextAreaElement).value = '0 0 0';
-      });
-    }
-
-    if (cancelBtn) {
-      cancelBtn.addEventListener('click', close);
-    }
-
-    if (applyBtn) {
-      applyBtn.addEventListener('click', () => {
-        const input = (dialog.querySelector('#rotation-center-input') as HTMLTextAreaElement).value;
-        const constraint = (
-          dialog.querySelector('input[name="center-constraint"]:checked') as HTMLInputElement
-        ).value;
-        const values = parseSpaceSeparatedValues(input);
-
-        if (values.length === 3) {
-          const [x, y, z] = values;
-
-          // Store current camera state
-          const currentPosition = this.camera.position.clone();
-          const currentQuaternion = this.camera.quaternion.clone();
-          const currentTarget = this.controls.target.clone();
-
-          // Set the new rotation center
-          this.controls.target.set(x, y, z);
-
-          // Apply constraint logic
-          if (constraint === 'rotation') {
-            // Keep rotation - restore quaternion and adjust position to maintain distance from new center
-            const distance = currentPosition.distanceTo(currentTarget);
-            this.camera.quaternion.copy(currentQuaternion);
-
-            // Position camera at distance from new center in same direction as rotation
-            const direction = new THREE.Vector3(0, 0, distance);
-            direction.applyQuaternion(currentQuaternion);
-            this.camera.position.copy(this.controls.target).add(direction);
-          } else {
-            // Keep position - restore position and adjust rotation to look at new center (default behavior)
-            this.camera.position.copy(currentPosition);
-            this.camera.lookAt(this.controls.target);
-          }
-
-          // Update controls and camera panel
-          this.controls.update();
-          this.updateCameraControlsPanel();
-
-          // Update axes position to show new rotation center
-          if ((this as any).axesHelper) {
-            (this as any).axesHelper.position.copy(this.controls.target);
-          }
-
-          console.log(
-            `🎯 Rotation center set to: (${x.toFixed(3)}, ${y.toFixed(3)}, ${z.toFixed(3)})`
-          );
-          this.updateRotationOriginButtonState();
-          close();
-        } else {
-          alert('Please enter exactly 3 numbers for center (X Y Z)');
-        }
-      });
-    }
+    transformDialogs.showRotationCenterDialog(this);
   }
 
   private openCalibrationFileDialog(fileIndex: number): void {
